Fixed Server delay option

+++ PRemote.Server/Program.cs
@@ -180,7 +180,7 @@ namespace PRemote.Server
                     switch (packet.SettingType)
                     {
                         case PDataType.Picture:
-                            await Task.Delay((int)packet.Data * 1000);
+                            await Task.Delay((byte)packet.Data * 1000);
                             await CameraList[i].CaptureImageAsync();
                             break;
                         case PDataType.Aperture:
